Inicio tests

+++ src/main/java/com/grupo08/socialmeli/controller/ProductController.java
@@ -25,7 +25,7 @@ public class ProductController {
         return new ResponseEntity<>(postService.getAll(),HttpStatus.OK);
     }
     @GetMapping("/followed/{userId}/list")
-    public ResponseEntity<?> followProcducts(@PathVariable("userId") Long idUser,@RequestParam(required = false)  String order){
+    public ResponseEntity<?> followProcducts(@PathVariable("userId") Integer idUser,@RequestParam(required = false)  String order){
         System.out.println("-------"+order+"--------");
         if(order==null){
             return new ResponseEntity<>(userService.postSortWeeks(idUser),HttpStatus.OK);
